// src/app/interfaces/new-user.interface.ts
export interface NewUser {
    username: string;
    password: string;
    confirmPassword: string;
    firstName: string;
    secondName: string;
    firstLastName: string;
    secondLastName: string;
    dpi: number;
    nit: string;
    phone: number;
    birthday: Date;
    country: string;
    state: string;
    city: string;
    complement: string;
}